requirements.txt
streamlit
SpeechRecognition
pyttsx3
google.generativeai
fpdf
PyAudio